Another enhancment to assure that an "prepend" mbuf is available. This also changes the copy boundary. User data is now kept with just enough prepended room for a data chunk header.
When we do the copy out, we now first
get a mbuf that is setup JUST for prepend.
Add it to the front.. and then copy out.

This way all drivers and lower level entities
will be able to prepend to the mbuf. Note
that I saw a 30Mbit per sec jump in a
2 mbuf chain

New transitions should now be at 204 and 436 bytes.


git-svn-id: http://sctp-refimpl.googlecode.com/svn/trunk@969 9df1edf5-d72c-5b5f-11c0-5f5209eb73f7