First of all finish up the mbcnt stuff. I forgot to add to the new accepted socket the mbuf cnt.
Second I now am attempting to loosen the locking
stratgey a bit. We no longer insist on always
having the INP lock owned before locking the TCB.
Basically we insist that if you have a TCB
lock, you never lock upwards to an INP. However
you can just lock a TCB directly without first
locking an INP .. for the few cases where you
would see

SCTP_INP_RLOCK(inp)
SCTP_TCB_LOCK(stcb);
SCTP_INP_RUNLOCK(inp);

in the code. I now have the INP locks ifdef'd
with sctp_invarients. This is a cautious approach
since we can easily back this out.. plus the
invarients code checks for this.. so we have
to still have it if we turn on the sctp_invarients.


git-svn-id: http://sctp-refimpl.googlecode.com/svn/trunk@1186 9df1edf5-d72c-5b5f-11c0-5f5209eb73f7