Get rid of create_lock. Now its a timer.
Found and fixed many lock inversions.

Still have one witness warning that I can't get rid of.. since
it is an invalid warning.. I am getting a lock
on the old and new socket buffer when moving
counts around. I have to lock both... I suppose
if I really wanted to I could lock one and do the
subtraction, then lock the other and do the
addition...  but that seems like a lot of work
for something that I am doing that is valid :-0


git-svn-id: http://sctp-refimpl.googlecode.com/svn/trunk@1263 9df1edf5-d72c-5b5f-11c0-5f5209eb73f7