Ok two issues.
In append_to_readqueue I was checking things BEFORE the
lock. This is bad since that allowed the reader to change
things on me ...causing a panic.

Now for the peeloff/accept.. we DID have the right function in
sctputil.c for moving the data to the new endpoint.. but it
had the old locking..i.e. used the socketbuf lock.

I have moved this to use the RIGHT lock.. but also there
is the issue of how to prevent a reader from accessing the
socket. So what we do is a sblock().. without wait. If it
fails we unlock and give up. The reality is the ONLY one
that will have this is someone doing a peeloff and
a seperate thread doing a read() at the same time. And in
such a case, we will leave the data stranded on the old
socket... since to do otherwise would put the system in
jeopardy.


git-svn-id: http://sctp-refimpl.googlecode.com/svn/trunk@1823 9df1edf5-d72c-5b5f-11c0-5f5209eb73f7