Two things.
It appears I had

<-----SHUTDOWN_ACK()
and
-->HB-Timer

Collied. The HB timer did a SCTP_TCB_LOCK(stcb); and then
block all further timers.

We do two things on this:
a) Stop all timers when we send a shutdown.. except the
   shutdown guard timer and shutdown timer.

and

b) Increment the refcnt on the assoc, before we try
   to lock.

now there is still a gap there.. but really stopping
the timers should fix the problem..

This is all related to the 1500 assoc's all dieing
at once.. with michaels run-away thread that started
all the assoc's.


git-svn-id: http://sctp-refimpl.googlecode.com/svn/trunk@1972 9df1edf5-d72c-5b5f-11c0-5f5209eb73f7