Another interesting race condition I found on the plane:
When you have have a true restart, there were
locks missing that meant if the sender was in the
process of sending a msg and was interupted it might
not be on the send queue for report_all_lost to find,
and then the TSN would be changed. This should not
effect MAC's though since they have a more global
lock. I wonder what the mpi guys are doing not
noticing notifications (comm lost etc). In TCP
they would get an error when the connection reset.
Without the notifications they don't know :-0


git-svn-id: http://sctp-refimpl.googlecode.com/svn/trunk@2317 9df1edf5-d72c-5b5f-11c0-5f5209eb73f7