This should fix both MAC and FreeBSD's crash inside sctp_inpcb_free() removing something from the sctp_hash.
There was a race that allowed a multi-processor
to wake up the acceptor, before the cookie processing
had finished for the TCP model. We now do the
sonewconn() without being connected. And then
at the end issue the soisconnected(). I also take
care of the problem that would occur if you had
more CPU's than the so_backlog. This would cause
the first in queue to get a soabort() which would
then hit our same issue (without a fix) inside
sctp_inpcb_free. We now mark them to UNBOUND
which will allow normal cleanup to work. As
well as increment the refcount.. so he wont
clean us up while we are linking it in.


git-svn-id: http://sctp-refimpl.googlecode.com/svn/trunk@2601 9df1edf5-d72c-5b5f-11c0-5f5209eb73f7