Ok we update the BSD timestamps with the commit I did today.. and put in place the VRF hash. Right now nothing gets into it but the single VRF entry..
Todo:

a) Build sctp IFN/IFA list
b) Convert init/initack code to use it
c) make asconf code use it and maintain it.
d) then finally attack the src address selection.

Note we still have an issue with DETACHED IPV6 addresses. It
looks like these get updated as the RA's fail to come in on
the flags. This means we will have to lock to check this
flag rather ugly... I had hoped we did not have to
deal with looking at either the IFA or IFN after
we built our list.. but I can't see how without
changing the V6 code to have a call back when something
goes detached.


git-svn-id: http://sctp-refimpl.googlecode.com/svn/trunk@2752 9df1edf5-d72c-5b5f-11c0-5f5209eb73f7