Ok it appears I have the iterator working fine with the new asconf code.
We COULD now go in and easily add per assoc asconf's
as well.. but that will require a special set of
iterators.

Also I need to look at the threading issues now and
how we migrate BSD over to threads.

But before that I need to test the singly homed case
and the subset bound case. And of course it would be
nice to test a bit more on src address selection.. but
I need to down stewart for this to rev it up to
a bit faster processor ;-) (note to all stewart
will be down for an hour or so sometime today).


git-svn-id: http://sctp-refimpl.googlecode.com/svn/trunk@2796 9df1edf5-d72c-5b5f-11c0-5f5209eb73f7