Consider the case of a Singly Homed (SH) Mobile node:
It changes address and sends
---asconf(ADD(B)/DEL(A)/SET-PRIM(B))->
Now the reciever cannot set primary. It would
refuse. But A which WAS the primary is deleted.

The code before this commit, would then just
set the primary to the very first net.. aka (B).
But this is unconfirmed.. very bad.

So what this does is:

a) Allow a set_primary() to be queued by placing
   a flag against the dest "REQ_PRIM". Then when
   HB-ac comes in, set the priamry. Note that if
   no HB-ACK before the dest hits failed (and still
   unconfimred) the flag is removed.

b) When the problem above with the SH node happens, we
   will ONLY allow CONTROL transmission in sender. This
   is not quite the spec, but I think its ok since
   I don't see an amplification attack one could do with
   control... at least not with our bundling.


git-svn-id: http://sctp-refimpl.googlecode.com/svn/trunk@2886 9df1edf5-d72c-5b5f-11c0-5f5209eb73f7