First get rid of printfs in threshold mngt that I put in for debug
Second it would appear that the KAME stack in the end
will not let you send
Global -> Linklocal

This makes it so this is no longer a perfered method
and only a acceptable method for src addr selection.

I need to go read the address scoping RFC/Draft and see
if this is bad behavior on KAME part or not.


git-svn-id: http://sctp-refimpl.googlecode.com/svn/trunk@2889 9df1edf5-d72c-5b5f-11c0-5f5209eb73f7