Fix for sctp_opt_info failures. structure sctp_paddrparams, sctp_paddrinfo, sctp_setprim, sctp_setpeerprim didnt have asoc id as the first member of the structure anymore which is expected by sctp_opt_info.
Move back asoc id as the first member. Add 4 byte padding for the structs to follow or move out the padding in the end to earlier.

Randy/Michael please double check the diffs from 64 bit angle as well. Other way to fix this would be to fix sctp_opt_info api itself not to assume this.


git-svn-id: http://sctp-refimpl.googlecode.com/svn/trunk@4128 9df1edf5-d72c-5b5f-11c0-5f5209eb73f7