Note sure if this compiles, but it implements Kozuka-san's proposed changes to make INIT-ACK's not use the source address selection address, and instead override the address to the destination address as the source. This will fix the case of  src=::1 dest=fe80::1 -> INIT
An abort will follow the INIT-ACK currently (before this commit) since
we won't list fe80::1 and we don't allow the src address to be
fe80::1 (private and loopback) in src address selection.

This may well cause a problem when we reply to an address, since
the route out of the box may be ingress filtered. Hmm probably should
fix this so that we use src-addr-selection if its not a loopback.


git-svn-id: http://sctp-refimpl.googlecode.com/svn/trunk@4914 9df1edf5-d72c-5b5f-11c0-5f5209eb73f7