When a send failure goes up we must ASSURE that
1) ssf_sinfo->sinfo_flags is set to the right FRAGMENT bits
2) We trim off the sctp data chunk headers if present.

This allows an application to see the REAL data and
to know if some has been take off the top.

That way one can look for the first Packet of
the message (if its not complete) and then do action
based on that.. a useful feature.

Before this commit we sometimes sent the chunk header
and other times (when it was still on the stream queue)
sent the data only... and worse never indicated if
stream queue data was chopped off i.e. some_take set

This left things confusing for an app.


git-svn-id: http://sctp-refimpl.googlecode.com/svn/trunk@5012 9df1edf5-d72c-5b5f-11c0-5f5209eb73f7