ID-sync
platforms other than bsd should check and see if
they can make the CONTENDED macro in the lock_h directory
work. We are now using contention on the lock as an
additional indication that there is a pending reference
to an inp...

If your locking system can tell you someone is waiting
on the lock you should put that test in instead of the (0)
I added.


git-svn-id: http://sctp-refimpl.googlecode.com/svn/trunk@6124 9df1edf5-d72c-5b5f-11c0-5f5209eb73f7